pyproject.toml
[project]
name = "local-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=0.34.0",
    "datasets>=2.20.0",
    "fastapi>=0.135.2",
    "google-genai>=1.68.0",
    "httpx>=0.28.1",
    "librosa>=0.11.0",
    "numpy>=2.4.3",
    "phonemizer>=3.3.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "scipy>=1.17.1",
    "sentencepiece>=0.2.1",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "torch>=2.11.0",
    "transformers>=5.4.0",
    "uvicorn>=0.42.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.8",
]
